fill Nitrogen back to 850
Assume 1860 is pre-industrial background N-deposition rates and copy
that value back to 0850 a.d.  This is obviously a flawed assumption,
but because there is a spatial pattern to the N-deposition (with a
hotspot in an unexpected area of MO), Mike, Dave, Yao, Jason & I
decided this was better than causing abrupt shifts in the N record.
For the most part, the 1860 N-deposition rates are much lower than
post-industrialization.