Change Biome driver base to Synmap
Based on discussions with the core PalEON modeling group, we decided to
move away from Ramankutty & Foley as the biome map to SYNMAP for two
reasons: 1) it’s what MsTMIP uses; 2) it natively provides a fraction
biome coverage.

To translate SYNMAP for PalEON, two products were made that are related
to each other, but do not translate directly:
1) Biome classification — this is for models like SiBCASA that assign a
single biome designation to an entire grid cell.
2) Fraction PFT coverage — this is for models like CLM or JULES that
prescript a PFT distribution that does not change.